requirements.txt
Flask==3.0.0
Flask-Caching==2.1.0
requests==2.31.0
python-dotenv==1.0.0

# Extracteur de factures dependencies
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
anthropic==0.39.0
Pillow==10.1.0
pdf2image==1.16.3
reportlab==4.0.7
werkzeug==3.0.1
